+++ conans/test/functional/command/export/export_test.py
@@ -45,8 +45,7 @@ class TestConan(ConanFile):
 
         client.save({"conanfile.py": GenConanfile()})
         client.run("export . lib/1.0@lasote/channel")
-        self.assertIn("lib/1.0@lasote/channel: A new conanfile.py version was exported",
-                      client.out)
+        self.assertIn("lib/1.0@lasote/channel: A new conanfile.py version was exported", client.out)
 
         client.save({"conanfile.py": GenConanfile("Lib", "1.0")})
         client.run("export . lasote", assert_error=True)
@@ -67,10 +66,9 @@ class TestConan(ConanFile):
         export_path = client.cache.package_layout(ref).export()
         export_src_path = client.cache.package_layout(ref).export_sources()
 
-        files = {CONANFILE: conanfile,
-                 "file1.txt": "",
-                 "file2.txt": ""}
-        client.save(files)
+        client.save({CONANFILE: conanfile,
+                     "file1.txt": "",
+                     "file2.txt": ""})
         mode1 = os.stat(os.path.join(client.current_folder, "file1.txt")).st_mode
         mode2 = os.stat(os.path.join(client.current_folder, "file2.txt")).st_mode
         os.chmod(os.path.join(client.current_folder, "file1.txt"), mode1 & ~stat.S_IWRITE)
@@ -84,12 +82,12 @@ class TestConan(ConanFile):
         with self.assertRaises(IOError):
             save(os.path.join(export_src_path, "file2.txt"), "")
         self.assertIn("WARN: Conanfile doesn't have 'license'", client.out)
-        files = {CONANFILE: conanfile,
-                 "file1.txt": "file1",
-                 "file2.txt": "file2"}
+
         os.chmod(os.path.join(client.current_folder, "file1.txt"), mode1 | stat.S_IWRITE)
         os.chmod(os.path.join(client.current_folder, "file2.txt"), mode2 | stat.S_IWRITE)
-        client.save(files)
+        client.save({CONANFILE: conanfile,
+                     "file1.txt": "file1",
+                     "file2.txt": "file2"})
         client.run("export . lasote/stable")
 
         self.assertEqual(load(os.path.join(export_path, "file1.txt")), "file1")
@@ -97,10 +95,9 @@ class TestConan(ConanFile):
         client.run("install Hello/1.2@lasote/stable --build=missing")
         self.assertIn("Hello/1.2@lasote/stable: Generating the package", client.out)
 
-        files = {CONANFILE: conanfile,
-                 "file1.txt": "",
-                 "file2.txt": ""}
-        client.save(files)
+        client.save({CONANFILE: conanfile,
+                     "file1.txt": "",
+                     "file2.txt": ""})
         os.chmod(os.path.join(client.current_folder, "file1.txt"), mode1 & ~stat.S_IWRITE)
         os.chmod(os.path.join(client.current_folder, "file2.txt"), mode2 & ~stat.S_IWRITE)
         client.run("export . lasote/stable")
@@ -110,8 +107,7 @@ class TestConan(ConanFile):
         self.assertIn("Hello/1.2@lasote/stable: Generating the package", client.out)
 
     def test_code_parent(self):
-        """ when referencing the parent, the relative folder "sibling" will be kept
-        """
+        # when referencing the parent, the relative folder "sibling" will be kept
         base = """
 from conans import ConanFile
 class TestConan(ConanFile):
@@ -121,9 +117,8 @@ class TestConan(ConanFile):
 """
         for conanfile in (base, base.replace("../*.txt", "../sibling*")):
             client = TestClient()
-            files = {"recipe/conanfile.py": conanfile,
-                     "sibling/file.txt": "Hello World!"}
-            client.save(files)
+            client.save({"recipe/conanfile.py": conanfile,
+                         "sibling/file.txt": "Hello World!"})
             client.current_folder = os.path.join(client.current_folder, "recipe")
             client.run("export . lasote/stable")
             ref = ConanFileReference("Hello", "1.2", "lasote", "stable")
@@ -154,18 +149,17 @@ class TestConan(ConanFile):
     def test_code_several_sibling(self):
         # if provided a path with slash, it will use as a export base
         client = TestClient()
-        conanfile = """
-from conans import ConanFile
-class TestConan(ConanFile):
-    name = "Hello"
-    version = "1.2"
-    exports_sources = "../test/src/*", "../cpp/*", "../include/*"
-"""
-        files = {"recipe/conanfile.py": conanfile,
-                 "test/src/file.txt": "Hello World!",
-                 "cpp/file.cpp": "Hello World!",
-                 "include/file.h": "Hello World!"}
-        client.save(files)
+        conanfile = textwrap.dedent("""
+            from conans import ConanFile
+            class TestConan(ConanFile):
+                name = "Hello"
+                version = "1.2"
+                exports_sources = "../test/src/*", "../cpp/*", "../include/*"
+            """)
+        client.save({"recipe/conanfile.py": conanfile,
+                     "test/src/file.txt": "Hello World!",
+                     "cpp/file.cpp": "Hello World!",
+                     "include/file.h": "Hello World!"})
         client.current_folder = os.path.join(client.current_folder, "recipe")
         client.run("export . lasote/stable")
         ref = ConanFileReference("Hello", "1.2", "lasote", "stable")
@@ -177,15 +171,14 @@ class TestConan(ConanFile):
     def test_filename(self, filename):
         client = TestClient()
         client.save({filename: GenConanfile("Hello", "1.2")})
-        client.run("export %s lasote/stable" % filename)
-        self.assertIn("Hello/1.2@lasote/stable: A new conanfile.py version was exported",
-                      client.out)
-        ref = ConanFileReference("Hello", "1.2", "lasote", "stable")
+        client.run("export %s user/stable" % filename)
+        self.assertIn("Hello/1.2@user/stable: A new conanfile.py version was exported", client.out)
+        ref = ConanFileReference("Hello", "1.2", "user", "stable")
         export_path = client.cache.package_layout(ref).export()
         conanfile = load(os.path.join(export_path, "conanfile.py"))
-        self.assertIn('name = "Hello"', conanfile)
+        self.assertIn("name = 'Hello'", conanfile)
         manifest = load(os.path.join(export_path, "conanmanifest.txt"))
-        self.assertIn('conanfile.py: cac514c81a0af0d87fa379b0bf16fbaa', manifest)
+        self.assertIn('conanfile.py: aa4e485651a4281b79d343af7d7b20e4', manifest)
 
     def test_exclude_basic(self):
         client = TestClient()
