+++ conans/test/functional/command/info/info_command_test.py
