+++ conans/test/functional/command/info/info_options_test.py
