+++ conans/test/functional/command/info/info_test.py
